clean wom lists
