+++ src/screens/WahlOMeter/ChartNote.js
@@ -5,8 +5,8 @@ const ChartNote = styled.Text`
   font-size: 13;
   color: #4a4a4a;
   text-align: center;
-  padding-horizontal: 18;
   padding-vertical: 18;
+  padding-horizontal: 18;
 `;
 
 export default ChartNote;
